Split glass-action into separate modules
