+++ glass-action/glass_action/main.py
@@ -1,0 +1,6 @@
+import click
+
+@click.group()
+@click.pass_context
+def main(ctx, **kw):
+    ctx.obj = {}
